+++ api/app.js
@@ -38,7 +38,19 @@ app.use(
 );
 
 app.use(helmet());
-app.use(express.json());
+
+app.use(
+  express.json({
+    // We need the raw body to verify webhook signatures.
+    // Let's compute it only when hitting the Stripe webhook endpoint.
+    verify: function (req, res, buf) {
+      if (req.originalUrl.startsWith('/webhook')) {
+        req.rawBody = buf.toString();
+      }
+    },
+  })
+);
+
 app.use(
   cors({
     origin: '*',
@@ -80,5 +92,8 @@ app.use(function (err, req, res, next) {
   }
   next(err);
 });
+
+
+
 
 module.exports = app;
